requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
backoff==2.2.1
colorama==0.4.6
distro==1.9.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
mcp==1.9.3
# openai==1.84.0
openai-agents==0.0.17
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.6
# starlette==0.47.0
tqdm==4.67.1
types-requests==2.32.0.20250602
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
langchain-groq>=0.3.2
langchain-community>=0.3.23
langgraph
fastapi==0.115.12
starlette==0.46.2
pinecone
# uvicorn==0.34.2
jiter==0.10.0
openai==1.82.1
tiktoken==0.9.0
langchain-core==0.3.63
langchain-openai==0.3.18
langsmith==0.3.43
boto3==1.38.36
botocore==1.38.36
jmespath==1.0.1
s3transfer==0.13.0
dnspython==2.7.0
pymongo==4.13.1
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aioitertools==0.12.0
boto3==1.38.36
boto3==1.38.27
botocore==1.38.36
botocore==1.38.27
wrapt==1.17.2